GameRenderCommands Class + GameTechRenderer inherits from it
GameRenderCommands is an abstract class stores the gameWorld, the activeObjects and the methods that will be needed by both OpenGL and PS4 to build the scene(Both need to make a skybox, sort list etc).

GameTechRenderer now inherits from OGLRenderer and GameRenderCommands. This means that it will use the base OpenGL renderer, and then have to implement the specific commands to draw the game in OpenGL.

Uses multiple inheritance, so not great. Research better way to do it if possible.