RendererBase parent of Renderer + No OpenGL debug drawing in Renderer
RendererBase was made to be a parent of renderer. This is a generic abstract class for all actual renderers. It provides base members such as a rendererAPI and functions for drawing debug data.

Renderer is now a child of RendererBase. Renderer extends the base class by adding a game world.

This was changed because Debug.h was causing lots of include errors when the Renderer class was included into it. This was because it was including GameWorld, which includes a lot of files that include Debug.h. To fix this error, this is why the RendererBase class was created.